requirements.txt
aioredis==2.0.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-limiter==0.1.6
greenlet==3.0.3
h11==0.14.0
idna==3.6
install==1.3.5
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
pip-autoremove==0.10.0
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.1.0b3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
starlette-context==0.3.6
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.0.post1
validators==0.22.0
